requirements.txt
claude-code-sdk>=0.0.14
fastapi>=0.115.0
uvicorn[standard]>=0.34.0
python-dotenv>=1.0.0
pydantic>=2.10.0
twilio>=9.0.0
python-multipart>=0.0.9
sqlalchemy[asyncio]>=2.0.0
asyncpg>=0.30.0
mcp>=1.0.0
apscheduler>=3.10.0
croniter>=2.0.0
